fourth

+++ src/main/java/com/manju/entity/Product.java
@@ -5,13 +5,17 @@ import javax.persistence.Entity;
 import javax.persistence.Id;
 import javax.persistence.Table;
 
+import org.springframework.stereotype.Component;
+
+
+
 @Entity
 @Table(name = "product")
 public class Product {
 
 	@Id
 	@Column(name = "pid")
-	private long pid;
+	private int pid;
 	
 	@Column(name = "product_name")
 	private String productName;
@@ -28,10 +32,10 @@ public class Product {
 	public String toString() {
 		return "Product {pid=" + pid + ", productName=" + productName + ", qty=" + qty + ", price=" + price + "}";
 	}
-	public long getPid() {
+	public int getPid() {
 		return pid;
 	}
-	public void setPid(long pid) {
+	public void setPid(int pid) {
 		this.pid = pid;
 	}
 	public String getProductName() {
@@ -52,7 +56,7 @@ public class Product {
 	public void setPrice(int price) {
 		this.price = price;
 	}
-	public Product(long pid, String productName, int qty, int price) {
+	public Product(int pid, String productName, int qty, int price) {
 		super();
 		this.pid = pid;
 		this.productName = productName;
